Partially adjusted code from Phobos repository. *Fixed my bug in YAAM formula.

+++ sfall/Modules/BurstMods.cpp
@@ -29,7 +29,7 @@ static long compute_spray_center_div;
 static long compute_spray_target_mult;
 static long compute_spray_target_div;
 
-static long _fastcall ComputeSpray(DWORD* roundsLeftOut,  DWORD* roundsRightOut, DWORD totalRounds, DWORD* roundsCenterOut) {
+static long _fastcall ComputeSpray(DWORD* roundsLeftOut, DWORD* roundsRightOut, DWORD totalRounds, DWORD* roundsCenterOut) {
 
 	// roundsCenter = totalRounds * mult / div
 	long result = totalRounds * compute_spray_center_mult;
@@ -39,7 +39,7 @@ static long _fastcall ComputeSpray(DWORD* roundsLeftOut,  DWORD* roundsRightOut,
 	if (roundsCenter == 0) roundsCenter = 1;
 	*roundsCenterOut = roundsCenter;
 
-	int roundsLeft = (totalRounds - roundsCenter) / 2;
+	long roundsLeft = (totalRounds - roundsCenter) / 2;
 	*roundsLeftOut = roundsLeft;
 
 	*roundsRightOut = totalRounds - roundsCenter - roundsLeft;
